mocha in CI
