+++ .github/workflows/mocha.yml
@@ -1,0 +1,28 @@
+name: CI
+
+on:
+  push:
+    branches:
+      - main
+  pull_request:
+    branches:
+      - main
+
+jobs:
+  test:
+    runs-on: ubuntu-latest
+
+    steps:
+    - name: Checkout code
+      uses: actions/checkout@v3
+
+    - name: Set up Node.js
+      uses: actions/setup-node@v3
+      with:
+        node-version: '16' # or whichever Node.js version you use
+
+    - name: Install dependencies
+      run: npm install
+
+    - name: Run Mocha tests
+      run: npx mocha # Assumes "npm test" runs mocha
